Remove redundant casts of literal indices (`\[([0-9]+) as libc::c_int as usize\]` => `[$1]`).
These are literals and indices must be `usize`,
so there can't be any type ambiguity with the casts removed.

+++ src/cdef_apply_tmpl_16.rs
@@ -973,10 +973,10 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
     let have_tt: libc::c_int = ((*(*f).c).n_tc > 1 as libc::c_int as libc::c_uint)
         as libc::c_int;
     let sb128: libc::c_int = (*(*f).seq_hdr).sb128;
-    let resize: libc::c_int = ((*(*f).frame_hdr).width[0 as libc::c_int as usize]
-        != (*(*f).frame_hdr).width[1 as libc::c_int as usize]) as libc::c_int;
-    let y_stride: ptrdiff_t = PXSTRIDE((*f).cur.stride[0 as libc::c_int as usize]);
-    let uv_stride: ptrdiff_t = PXSTRIDE((*f).cur.stride[1 as libc::c_int as usize]);
+    let resize: libc::c_int = ((*(*f).frame_hdr).width[0]
+        != (*(*f).frame_hdr).width[1]) as libc::c_int;
+    let y_stride: ptrdiff_t = PXSTRIDE((*f).cur.stride[0]);
+    let uv_stride: ptrdiff_t = PXSTRIDE((*f).cur.stride[1]);
     let mut bit: libc::c_int = 0 as libc::c_int;
     let mut by: libc::c_int = by_start;
     while by < by_end {
@@ -999,7 +999,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                 ((*f)
                     .lf
                     .cdef_line[(tf == 0) as libc::c_int
-                    as usize][0 as libc::c_int as usize])
+                    as usize][0])
                     .offset(
                         ((have_tt * sby * 4 as libc::c_int) as isize * y_stride)
                             as isize,
@@ -1007,7 +1007,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                 ((*f)
                     .lf
                     .cdef_line[(tf == 0) as libc::c_int
-                    as usize][1 as libc::c_int as usize])
+                    as usize][1])
                     .offset(
                         ((have_tt * sby * 8 as libc::c_int) as isize * uv_stride)
                             as isize,
@@ -1015,7 +1015,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                 ((*f)
                     .lf
                     .cdef_line[(tf == 0) as libc::c_int
-                    as usize][2 as libc::c_int as usize])
+                    as usize][2])
                     .offset(
                         ((have_tt * sby * 8 as libc::c_int) as isize * uv_stride)
                             as isize,
@@ -1030,9 +1030,9 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
         }
         let mut lr_bak: [[[[pixel; 2]; 8]; 3]; 2] = [[[[0; 2]; 8]; 3]; 2];
         let mut iptrs: [*mut pixel; 3] = [
-            ptrs[0 as libc::c_int as usize],
-            ptrs[1 as libc::c_int as usize],
-            ptrs[2 as libc::c_int as usize],
+            ptrs[0],
+            ptrs[1],
+            ptrs[2],
         ];
         edges = ::core::mem::transmute::<
             libc::c_uint,
@@ -1071,10 +1071,10 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                     .as_ptr()
                     .offset(by_idx as isize) as *const [uint16_t; 2];
                 noskip_mask = ((*noskip_row
-                    .offset(0 as libc::c_int as isize))[1 as libc::c_int as usize]
+                    .offset(0 as libc::c_int as isize))[1]
                     as libc::c_uint) << 16 as libc::c_int
                     | (*noskip_row
-                        .offset(0 as libc::c_int as isize))[0 as libc::c_int as usize]
+                        .offset(0 as libc::c_int as isize))[0]
                         as libc::c_uint;
                 y_lvl = (*(*f).frame_hdr).cdef.y_strength[cdef_idx as usize];
                 uv_lvl = (*(*f).frame_hdr).cdef.uv_strength[cdef_idx as usize];
@@ -1090,9 +1090,9 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                 uv_sec_lvl += (uv_sec_lvl == 3 as libc::c_int) as libc::c_int;
                 uv_sec_lvl <<= bitdepth_min_8;
                 bptrs = [
-                    iptrs[0 as libc::c_int as usize],
-                    iptrs[1 as libc::c_int as usize],
-                    iptrs[2 as libc::c_int as usize],
+                    iptrs[0],
+                    iptrs[1],
+                    iptrs[2],
                 ];
                 let mut bx: libc::c_int = sbx * sbsz;
                 while bx < imin((sbx + 1 as libc::c_int) * sbsz, (*f).bw) {
@@ -1157,8 +1157,8 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                 .expect(
                                     "non-null function pointer",
                                 )(
-                                bptrs[0 as libc::c_int as usize],
-                                (*f).cur.stride[0 as libc::c_int as usize],
+                                bptrs[0],
+                                (*f).cur.stride[0],
                                 &mut variance,
                                 (*f).bitdepth_max,
                             );
@@ -1186,7 +1186,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                     .offset(0 as libc::c_int as isize))
                                     .offset(offset as isize) as *mut pixel;
                             }
-                            bot = (bptrs[0 as libc::c_int as usize])
+                            bot = (bptrs[0])
                                 .offset(
                                     (8 * y_stride) as isize,
                                 );
@@ -1236,7 +1236,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                         (have_tt as isize * offset
                                             + (bx * 4 as libc::c_int) as isize) as isize,
                                     ) as *mut pixel;
-                                bot = (bptrs[0 as libc::c_int as usize])
+                                bot = (bptrs[0])
                                     .offset(
                                         (8 * y_stride) as isize,
                                     );
@@ -1249,13 +1249,13 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                 variance,
                             );
                             if adj_y_pri_lvl != 0 || y_sec_lvl != 0 {
-                                ((*dsp).cdef.fb[0 as libc::c_int as usize])
+                                ((*dsp).cdef.fb[0])
                                     .expect(
                                         "non-null function pointer",
                                     )(
-                                    bptrs[0 as libc::c_int as usize],
-                                    (*f).cur.stride[0 as libc::c_int as usize],
-                                    (lr_bak[bit as usize][0 as libc::c_int as usize])
+                                    bptrs[0],
+                                    (*f).cur.stride[0],
+                                    (lr_bak[bit as usize][0])
                                         .as_mut_ptr() as const_left_pixel_row_2px,
                                     top,
                                     bot,
@@ -1268,13 +1268,13 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                 );
                             }
                         } else if y_sec_lvl != 0 {
-                            ((*dsp).cdef.fb[0 as libc::c_int as usize])
+                            ((*dsp).cdef.fb[0])
                                 .expect(
                                     "non-null function pointer",
                                 )(
-                                bptrs[0 as libc::c_int as usize],
-                                (*f).cur.stride[0 as libc::c_int as usize],
-                                (lr_bak[bit as usize][0 as libc::c_int as usize])
+                                bptrs[0],
+                                (*f).cur.stride[0],
+                                (lr_bak[bit as usize][0])
                                     .as_mut_ptr() as const_left_pixel_row_2px,
                                 top,
                                 bot,
@@ -1388,7 +1388,7 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                                         "non-null function pointer",
                                     )(
                                     bptrs[pl as usize],
-                                    (*f).cur.stride[1 as libc::c_int as usize],
+                                    (*f).cur.stride[1],
                                     (lr_bak[bit as usize][pl as usize]).as_mut_ptr()
                                         as const_left_pixel_row_2px,
                                     top,
@@ -1407,13 +1407,13 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                         last_skip = 0 as libc::c_int;
                     }
                     bptrs[0 as libc::c_int
-                        as usize] = (bptrs[0 as libc::c_int as usize])
+                        as usize] = (bptrs[0])
                         .offset(8 as libc::c_int as isize);
                     bptrs[1 as libc::c_int
-                        as usize] = (bptrs[1 as libc::c_int as usize])
+                        as usize] = (bptrs[1])
                         .offset((8 as libc::c_int >> ss_hor) as isize);
                     bptrs[2 as libc::c_int
-                        as usize] = (bptrs[2 as libc::c_int as usize])
+                        as usize] = (bptrs[2])
                         .offset((8 as libc::c_int >> ss_hor) as isize);
                     bx += 2 as libc::c_int;
                     edges = ::core::mem::transmute::<
@@ -1426,13 +1426,13 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
                 }
             }
             iptrs[0 as libc::c_int
-                as usize] = (iptrs[0 as libc::c_int as usize])
+                as usize] = (iptrs[0])
                 .offset((sbsz * 4 as libc::c_int) as isize);
             iptrs[1 as libc::c_int
-                as usize] = (iptrs[1 as libc::c_int as usize])
+                as usize] = (iptrs[1])
                 .offset((sbsz * 4 as libc::c_int >> ss_hor) as isize);
             iptrs[2 as libc::c_int
-                as usize] = (iptrs[2 as libc::c_int as usize])
+                as usize] = (iptrs[2])
                 .offset((sbsz * 4 as libc::c_int >> ss_hor) as isize);
             sbx += 1;
             edges = ::core::mem::transmute::<
@@ -1441,23 +1441,23 @@ pub unsafe extern "C" fn dav1d_cdef_brow_16bpc(
             >(edges as libc::c_uint | CDEF_HAVE_LEFT as libc::c_int as libc::c_uint);
         }
         ptrs[0 as libc::c_int
-            as usize] = (ptrs[0 as libc::c_int as usize])
+            as usize] = (ptrs[0])
             .offset(
                 (8
-                    * PXSTRIDE((*f).cur.stride[0 as libc::c_int as usize])) as isize,
+                    * PXSTRIDE((*f).cur.stride[0])) as isize,
             );
         ptrs[1 as libc::c_int
-            as usize] = (ptrs[1 as libc::c_int as usize])
+            as usize] = (ptrs[1])
             .offset(
                 (8
-                    * PXSTRIDE((*f).cur.stride[1 as libc::c_int as usize]) >> ss_ver)
+                    * PXSTRIDE((*f).cur.stride[1]) >> ss_ver)
                     as isize,
             );
         ptrs[2 as libc::c_int
-            as usize] = (ptrs[2 as libc::c_int as usize])
+            as usize] = (ptrs[2])
             .offset(
                 (8
-                    * PXSTRIDE((*f).cur.stride[1 as libc::c_int as usize]) >> ss_ver)
+                    * PXSTRIDE((*f).cur.stride[1]) >> ss_ver)
                     as isize,
             );
         (*tc).top_pre_cdef_toggle ^= 1 as libc::c_int;
